pyproject.toml
[project]
name = "subtitle-tool"
version = "0.1.0"
description = "Tool for generating or synchronizing subtitles with help from Google Gemini AI."
authors = [{ name = "J. Eduardo", email = "j.eduardo@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.1",
    "click>=8.2.1",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.16.1",
    "humanize>=4.12.3",
    "pydub>=0.25.1",
    "pysubs2>=1.8.0",
    "tenacity>=9.1.2",
]

[project.scripts]
subtitle-tool = "subtitle_tool.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
